requirements.txt
# Deep Learning Framework
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0.74

# NLP & Text Processing
nltk>=3.8.1
spacy>=3.6.0

# Data Processing
numpy>=1.24.3
pandas>=2.0.3
h5py>=3.9.0
scikit-learn>=1.3.0

# FastAPI & Web Framework
fastapi>=0.103.0
uvicorn[standard]>=0.23.2
python-multipart>=0.0.6
pydantic>=2.3.0

# Evaluation Metrics
pycocoevalcap>=1.2
sacrebleu>=2.3.1

# Utilities
tqdm>=4.66.1
matplotlib>=3.7.2
seaborn>=0.12.2
tensorboard>=2.14.0
pyyaml>=6.0.1

# Development Tools
jupyter>=1.0.0
ipython>=8.14.0
pytest>=7.4.0
